+++ docs/src/components/appropriate-image/appropriate-image.module.scss
