+++ docs/src/components/appropriate-image/appropriate-image.tsx
