+++ docs/src/components/specs/render-signature.ts
