+++ docs/src/routes/deprecations.tsx
@@ -2,7 +2,7 @@ import {Markdown} from '~/components/markdown';
 import Caption from '~/components/caption';
 import SDKSupportTable from '~/components/sdk-support-table/sdk-support-table';
 import Property from '~/components/property';
-import Subtitle from '~/components/style-spec/specs/subtitle';
+import Subtitle from '~/components/specs/subtitle';
 function Layers() {
 
     return (
